requirements.txt
python-dotenv
flask_sqlalchemy
flask_bcrypt
flask_migrate
paho-mqtt